requirements_working.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
attrs==25.4.0
backoff==2.2.1
banks==2.2.0
bcrypt==5.0.0
beautifulsoup4==4.13.5
blinker==1.9.0
brotli==1.2.0
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
crewai==1.7.0
crewai-tools==1.7.0
cryptography==46.0.3
dataclasses-json==0.6.7
ddgs==9.9.3
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
durationpy==0.10
et_xmlfile==2.0.0
fake-useragent==2.2.0
fastuuid==0.14.0
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
griffe==1.15.0
grpcio==1.67.1
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.12.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.10.0
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
lancedb==0.5.7
langchain==1.2.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.1
langchain-ollama==1.0.0
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.56
litellm==1.80.0
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.3
llama-index-cli==0.5.1
llama-index-core==0.14.10
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-ollama==0.9.0
llama-index-llms-openai==0.5.6
llama-index-readers-file==0.5.5
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-chroma==0.5.3
llama-index-workflows==2.11.5
llama-parse==0.6.54
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mcp==1.16.0
mcpadapt==0.1.19
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.13.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
nodeenv==1.9.1
numpy==2.3.3
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.23.2
openai==1.83.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.11.5
ormsgpack==1.12.1
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
platformdirs==4.5.1
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.5.0
primp==0.15.0
propcache==0.4.1
protobuf==5.29.5
py==1.11.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.9.0
pylance==0.9.18
PyMuPDF==1.26.6
pypdf==6.4.1
pypdfium2==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytube==15.0.0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
ratelimiter==1.2.0.post0
referencing==0.37.0
regex==2024.9.11
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
semver==3.0.4
sentence-transformers==5.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.50.0
streamlit==1.52.1
striprtf==0.0.26
sympy==1.13.1
tavily-python==0.7.14
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.2
tomli_w==1.1.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.2
tqdm==4.67.1
transformers==4.46.3
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uuid_utils==0.12.0
uv==0.9.16
uvicorn==0.38.0
virtualenv==20.35.4
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
youtube-transcript-api==1.2.3
zipp==3.23.0
zstandard==0.25.0
